fix(vm): fix pop_f64 sign extension corrupting f64 values with negative low bits
When the low 32 bits of an f64 had bit 31 set, `pop_i32() as u64`
sign-extended to 64 bits, corrupting the upper 32 bits during the OR
reconstruction. Cast through u32 first to zero-extend.

Also clean up a stale vm_debug! and fix a broken if-let in CALL_NAT.

+++ crates/auto-lang/src/vm/virt_memory.rs
@@ -273,7 +273,7 @@ impl VirtualRAM {
     #[inline(always)]
     pub fn pop_f64(&mut self) -> f64 {
         let high = self.pop_i32() as u64;
-        let low = self.pop_i32() as u64;
+        let low = (self.pop_i32() as u32) as u64;
         let bits = (high << 32) | low;
         unsafe { f64::from_bits(bits) }
     }
